♻️ refactor: 로고 위치를 헤더에서 사이드바로 변경
