+++ src/components/common/header/Header.jsx
@@ -1,30 +1,15 @@
 import React, { useState } from "react";
 import {
   StyledHeaderDiv,
-  StyledHeaderLogoDiv,
   StyledHeaderInput,
   StyledHeaderParentDiv,
   StyledHeaderInputDiv,
   StyledSearchIcon,
-  StyledHeaderMoonIcon,
-  StyledHeaderSunIcon,
-  StyledHeaderFillMoonIcon,
-  StyledHeaderFillSunIcon,
 } from "./Header.style";
-import Logo from "../logo/Logo";
 
 export default function Header(props) {
-  const [isDarkMode, setIsDarkMode] = useState(false);
-
-  const toggleTheme = () => {
-    setIsDarkMode(!isDarkMode);
-  };
-
   return (
     <StyledHeaderDiv>
-      <StyledHeaderLogoDiv>
-        <Logo size={"m"} />
-      </StyledHeaderLogoDiv>
       <StyledHeaderParentDiv>
         <StyledHeaderInputDiv>
           <StyledHeaderInput
@@ -34,19 +19,6 @@ export default function Header(props) {
           />
           <StyledSearchIcon />
         </StyledHeaderInputDiv>
-        <div>
-          {isDarkMode ? (
-            <>
-              <StyledHeaderFillMoonIcon onClick={toggleTheme} filled />
-              <StyledHeaderSunIcon onClick={toggleTheme} />
-            </>
-          ) : (
-            <>
-              <StyledHeaderMoonIcon onClick={toggleTheme} />
-              <StyledHeaderFillSunIcon onClick={toggleTheme} filled />
-            </>
-          )}
-        </div>
       </StyledHeaderParentDiv>
     </StyledHeaderDiv>
   );
