+++ src/pages/main/Main.style.js
@@ -1,6 +1,7 @@
 import styled from "styled-components";
 
 export const StyledMainDiv = styled.div`
+  display: flex;
   width: 100%;
   height: 100%;
 `;
